requirements.txt
django==3.2.6
djangorestframework==3.12.4
django_extensions==3.1.3
psycopg2-binary==2.9.1
gunicorn==20.1.0
model_bakery==1.3.2
python-dateutil==2.8.2
drf-yasg==1.20.0